Students and non-editors don't need space for add assignment button
The reason we moved to “invisible button” instead of “no button” is so
that the table size wouldn’t change as the mouse moved. But in these
two cases, a button will _never_ appear; as such, we can save screen
real-estate buy just not showing the button.

(Sorry I didn’t create an issue here — this was just a quick thing I
noticed while trying the “View as” feature in student view!)